Split core builtins
