+++ Makefile
@@ -3,7 +3,7 @@ CFLAGS ?= -Wall -Wextra -std=c99
 PREFIX ?= /usr/local
 MANPREFIX ?= $(PREFIX)/share/man
 
-SRCS := src/builtins.c src/builtins_fs.c src/builtins_jobs.c \
+SRCS := src/builtins.c src/builtins_core.c src/builtins_fs.c src/builtins_jobs.c \
        src/builtins_alias.c src/builtins_func.c src/builtins_vars.c \
        src/builtins_read.c src/builtins_getopts.c src/vars.c \
        src/builtins_misc.c src/builtins_history.c src/builtins_time.c \
