+++ src/builtins.c
@@ -15,10 +15,6 @@ extern int builtin_cd(char **);
 extern int builtin_pushd(char **);
 extern int builtin_popd(char **);
 extern int builtin_dirs(char **);
-extern int builtin_exit(char **);
-extern int builtin_colon(char **);
-extern int builtin_true(char **);
-extern int builtin_false(char **);
 extern int builtin_echo(char **);
 extern int builtin_pwd(char **);
 extern int builtin_jobs(char **);
