+++ src/builtins_misc.c
@@ -33,52 +33,6 @@
 extern int last_status;
 extern FILE *parse_input;
 #include <sys/wait.h>
-extern void run_exit_trap(void);
-
-/* Exit the shell, freeing resources and using the provided status
- * or the status of the last command when none is given. */
-int builtin_exit(char **args) {
-    int status = last_status;
-    if (args[1]) {
-        char *end;
-        errno = 0;
-        long val = strtol(args[1], &end, 10);
-        if (*end != '\0' || errno != 0) {
-            fprintf(stderr, "usage: exit [STATUS]\n");
-            return 1;
-        }
-        status = (int)val;
-    }
-    run_exit_trap();
-    free_aliases();
-    free_functions();
-    free_shell_vars();
-    exit(status);
-}
-
-/* No-op builtin that always succeeds. */
-int builtin_colon(char **args)
-{
-    (void)args;
-    last_status = 0;
-    return 1;
-}
-
-/* Always succeed and set status to 0. */
-int builtin_true(char **args)
-{
-    (void)args;
-    last_status = 0;
-    return 1;
-}
-
-/* Always fail and set status to 1. */
-int builtin_false(char **args)
-{
-    (void)args;
-    last_status = 1;
-    return 1;
-}
 
 /* Print arguments separated by spaces. Supports -n to suppress the
  * trailing newline and -e to interpret common backslash escapes. */
